# service/apigateway/docs/swagger.yaml
definitions:
...
  response.ApiResponseCard:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/response.CardResponse'
        description: Card data payload
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status ("success" or "error")
        type: string
    type: object
...
  response.ApiResponsePaginationUser:
    properties:
      data:
        description: Array of user data
        items:
          $ref: '#/definitions/response.UserResponse'
        type: array
      message:
        description: Response message
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: Pagination metadata
      status:
        description: Response status
        type: string
    type: object
  response.ApiResponsePaginationUserDeleteAt:
    properties:
      data:
        description: Array of user data with deletion info
        items:
          $ref: '#/definitions/response.UserResponseDeleteAt'
        type: array
      message:
        description: Response message
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: Pagination metadata
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseRole:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/response.RoleResponse'
        description: Role data payload
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseUser:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/response.UserResponse'
        description: Contains the user data
      message:
        description: Descriptive response message
        type: string
      status:
        description: Response status ("success" or "error")
        type: string
    type: object
...
  response.CardResponse:
    properties:
      card_number:
        description: Masked card number (e.g., "4242********4242")
        type: string
      card_provider:
        description: Card issuer (e.g., "Visa", "MasterCard")
        type: string
      card_type:
        description: Type of card ("credit" or "debit")
        type: string
      created_at:
        description: Timestamp when card was added (RFC3339 format)
        type: string
      cvv:
        description: Masked CVV (typically "***")
        type: string
      expire_date:
        description: Card expiration date (MM/YY format)
        type: string
      id:
        description: Unique card identifier
        type: integer
      updated_at:
        description: Timestamp when card was last updated
        type: string
      user_id:
        description: ID of the card owner
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        description: HTTP status code or application error code
        type: integer
      message:
        description: Human-readable error message
        type: string
      status:
        description: Error status ("error", "fail", etc.)
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        description: Current page number (1-based index)
        type: integer
      page_size:
        description: Number of items per page
        type: integer
      total_pages:
        description: Total number of pages available
        type: integer
      total_records:
        description: Total number of items across all pages
        type: integer
    type: object
  response.RoleResponse:
    properties:
      created_at:
        description: Timestamp when role was created (RFC3339 format)
        type: string
      id:
        description: Unique role identifier
        type: integer
      name:
        description: Name of the role (e.g., "admin", "user")
        type: string
      updated_at:
        description: Timestamp when role was last updated
        type: string
    type: object
...
  response.UserResponse:
    properties:
      created_at:
        description: Timestamp when the user account was created
        type: string
      email:
        description: User's email address (used for authentication)
        type: string
      firstname:
        description: User's first name
        type: string
      id:
        description: Unique user identifier
        type: integer
      is_verified:
        description: Whether the user's email has been verified
        type: boolean
      lastname:
        description: User's last name
        type: string
      updated_at:
        description: Timestamp when the user account was last updated
        type: string
    type: object
  response.UserResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        description: Timestamp when user was soft deleted (nil if active)
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        description: Embedded UserResponse fields
        type: integer
      is_verified:
        type: boolean
      lastname:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/card-query/user:
    get:
      consumes:
      - application/json
      description: Retrieve a list of cards associated with a user by their ID
      produces:
      - application/json
      responses:
        "200":
          description: Card data
          schema:
            $ref: '#/definitions/response.ApiResponseCard'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve card record
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve cards by user ID
      tags:
      - Card Query
...
  /api/role-query/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a role by the associated user ID.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role data
          schema:
            $ref: '#/definitions/response.ApiResponseRole'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to fetch role by user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get role by user ID
      tags:
      - Role Command
...
  /api/user-query:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationUser'
        "500":
          description: Failed to retrieve user data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all users
      tags:
      - User Query
  /api/user-query/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/response.ApiResponseUser'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve user data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find user by ID
      tags:
      - User Query
  /api/user-query/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active users
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationUserDeleteAt'
        "500":
          description: Failed to retrieve user data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active users
      tags:
      - User Query
  /api/user-query/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed user records
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed user data
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationUserDeleteAt'
        "500":
          description: Failed to retrieve user data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed users
      tags:
      - User Query
...
swagger: "2.0"